Add files via upload
Eco evolutionary model with initially an abundance of floatting plants and absence of submerged (rooted) macrophytes. Our goal is to see for which values of the trait (zm) a new mutant (or migrant) of a submerged plant can grow depending on nutrient parameters to either coexist with F or outcompete them. The second step is to see when the apadtive dynamics on Sm would allow going deeper in the lake and help killing F in case of coexistance .